requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
yt-dlp==2024.1.1
openai-whisper==20240314
torch==2.1.2
numpy==1.26.3
soundfile==0.12.1
librosa==0.10.0
